task_2_DoubleLinkedList: Сформирован class DoubleLinkedList
