+++ Node_and_DoubleLinkedNode.py
@@ -31,6 +31,28 @@ class Node:
         self.is_valid(next_)
         self._next = next_
 
+
+class DoubleLinkedNode(Node):
+    def __init__(self, value: Any, prev: Optional["Node"] = None, next_: Optional["Node"] = None):
+        super().__init__(value, next_)
+        self.prev = prev
+
+    @property
+    def prev(self):
+        return self._prev  # TODO объект теперь вызываемый
+
+    @prev.setter
+    def prev(self, prev: Optional["Node"]):
+        self.is_valid(prev)
+        self._prev = prev  # TODO сделать слабую ссылку
+
+    def __repr__(self) -> str:
+        next_prev = None if self.prev is None else f"DoubleLinkedNode({self.prev})"
+        next_repr = None if self.next is None else f"DoubleLinkedNode({self.next})"  # todo make all
+
+        return f"DoubleLinkedNode({self.value}, {next_prev}, {next_repr})"
+
+
 if __name__ == '__main__':
     """
     Тестируем класс Node
